refactor(evm-wallet): extract shared tx helpers to lib/tx-utils
Move applyGasBuffer, validateGasEstimate, and validateTokenCallResult
out of both coordinator vats into src/lib/tx-utils.ts. Both vats now
import from the shared module.
